requirements.txt
Adafruit-Blinka==8.39.1
adafruit-circuitpython-busdevice==5.2.9
adafruit-circuitpython-connectionmanager==3.1.0
adafruit-circuitpython-debouncer==2.0.8
adafruit-circuitpython-requests==4.0.0
adafruit-circuitpython-ticks==1.0.13
adafruit-circuitpython-typing==1.10.3
Adafruit-PlatformDetect==3.63.0
Adafruit-PureIO==1.1.11
pimoroni-ioexpander==0.0.5
pyftdi==0.55.4
pyserial==3.5
pyusb==1.2.1
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
smbus2==0.4.3
sysv-ipc==1.1.0
typing_extensions==4.11.0
